Refactor endre periode functionality by consolidating modal steps into a panel structure
- Removed EndrePeriodeModal and VelgPeriodeModalStep components.
- Introduced EndrePeriodePanel to manage the period modification process.
- Updated form handling and state management for selecting and editing periods.
- Adjusted imports and references throughout the codebase to align with the new structure.
- Updated localization keys from "endrePeriodeModal" to "endrePeriodePanel" for consistency.
