+++ packages/ny-uttaksplan/src/components/spørsmål/GraderingSpørsmål.tsx
@@ -7,13 +7,13 @@ import { StønadskontoType } from '@navikt/fp-constants';
 import { RhfNumericField, RhfRadioGroup } from '@navikt/fp-form-hooks';
 import { isRequired } from '@navikt/fp-validation';
 
-import { EndrePeriodeModalStepFormValues } from '../endre-periode-modal/steps/EndrePeriodeModalStep';
+import { EndrePeriodePanelStepFormValues } from '../endre-periode-panel/steps/EndrePeriodePanelStep';
 import { LeggTilPeriodePanelFormValues } from '../legg-til-periode-panel/types/LeggTilPeriodePanelFormValues';
 import { prosentValideringGradering } from './validators';
 
 export const GraderingSpørsmål = () => {
     const intl = useIntl();
-    const { watch, control } = useFormContext<LeggTilPeriodePanelFormValues | EndrePeriodeModalStepFormValues>();
+    const { watch, control } = useFormContext<LeggTilPeriodePanelFormValues | EndrePeriodePanelStepFormValues>();
 
     const graderingValue = watch('skalDuJobbe');
     const kontoTypeValue = watch('kontoType');
