+++ packages/ny-uttaksplan/src/components/spørsmål/SamtidigUttakSpørsmål.tsx
@@ -6,12 +6,12 @@ import { Radio, VStack } from '@navikt/ds-react';
 import { RhfNumericField, RhfRadioGroup } from '@navikt/fp-form-hooks';
 import { isRequired } from '@navikt/fp-validation';
 
-import { EndrePeriodeModalStepFormValues } from '../endre-periode-modal/steps/EndrePeriodeModalStep';
+import { EndrePeriodePanelStepFormValues } from '../endre-periode-panel/steps/EndrePeriodePanelStep';
 import { LeggTilPeriodePanelFormValues } from '../legg-til-periode-panel/types/LeggTilPeriodePanelFormValues';
 import { valideringSamtidigUttak } from './validators';
 
 export const SamtidigUttakSpørsmål = () => {
-    const { watch, control } = useFormContext<LeggTilPeriodePanelFormValues | EndrePeriodeModalStepFormValues>();
+    const { watch, control } = useFormContext<LeggTilPeriodePanelFormValues | EndrePeriodePanelStepFormValues>();
 
     const samtidigUttakValue = watch('samtidigUttak');
     const stillingsprosentValue = watch('stillingsprosent');
