+++ packages/ny-uttaksplan/src/intl/en_US.json
@@ -139,7 +139,7 @@
     "leggTilPeriodePanel.samtidiguttaksprosent.måVæreEtTall": "The percentage must be a number",
     "leggTilPeriodePanel.samtidiguttaksprosent.måVæreStørreEnn0": "The percentage must be a number greater than 0",
     "leggTilPeriodePanel.samtidiguttaksprosent.måVæreMindreEnn100": "The percentage must be a number less than or equal to 100",
-    "endrePeriodeModal.tittel": "Change period",
+    "endrePeriodePanel.tittel": "Change period",
     "endreTidsPeriodeModal.fom.påkrevd": "You must specify a start date",
     "endreTidsPeriodeModal.fom.gyldigDato": "The start date must be a valid date in dd.mm.yyyy format",
     "endreTidsPeriodeModal.fom.førTilDato": "The start date must be equal to or before the end date.",
