+++ packages/ny-uttaksplan/src/intl/nb_NO.json
@@ -139,7 +139,7 @@
     "leggTilPeriodePanel.samtidiguttaksprosent.måVæreEtTall": "Prosenten må være et tall",
     "leggTilPeriodePanel.samtidiguttaksprosent.måVæreStørreEnn0": "Prosenten må være et tall større enn 0",
     "leggTilPeriodePanel.samtidiguttaksprosent.måVæreMindreEnn100": "Prosenten må være et tall mindre eller lik 100",
-    "endrePeriodeModal.tittel": "Endre periode",
+    "endrePeriodePanel.tittel": "Endre periode",
     "endreTidsPeriodeModal.fom.påkrevd": "Du må oppgi en startdato",
     "endreTidsPeriodeModal.fom.gyldigDato": "Startdatoen må være en gyldig dato på formatet dd.mm.åååå",
     "endreTidsPeriodeModal.fom.førTilDato": "Startdatoen må være lik eller før sluttdatoen.",
